Python/Goods_Matcher/requirements.txt
streamlit
python-dotenv
pyodbc
fuzzywuzzy
python-Levenshtein
jellyfish
phonetics
pandas

